requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pymupdf
pydantic
python-dotenv
numpy
orjson
openai
python-dotenv